API to upload image on cloudinary

+++ Backend/src/utils/cloudinary.js
@@ -1,6 +1,6 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
-
+import path from "path";
 // fs is file System
 
 cloudinary.config({
@@ -8,6 +8,11 @@ cloudinary.config({
   api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
+
+const isValidFilePath = (filePath) => {
+  // Check if file path is an actual path and not a data URI
+  return path.isAbsolute(filePath) && !filePath.startsWith('data:');
+};
 
 const uploadOnCloudinary = async (localFilePath) => {
   try {
@@ -20,10 +25,14 @@ const uploadOnCloudinary = async (localFilePath) => {
 
     // File has been uploaded Successfully
     console.log("File is uploaded Successfully on cloudinary", response.url);
-    fs.unlinkSync(localFilePath);
+    // fs.unlinkSync(localFilePath);
+    if (isValidFilePath(localFilePath)) fs.unlinkSync(localFilePath);
+
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // Remove the locally saved temporary file as the upload operation got failed
+    // fs.unlinkSync(localFilePath); // Remove the locally saved temporary file as the upload operation got failed
+    if (isValidFilePath(localFilePath)) fs.unlinkSync(localFilePath);
+
     return null;
   }
 };
